refactor: additional imports adjustments
