+++ sea-of-code/src/pages/auth/components/OAuthButtons.tsx
@@ -1,6 +1,6 @@
 import Button from '../../../components/ui/Button';
-import gitHubLogo from '../../assets/gitHub-logo-black.svg';
-import googleLogo from '../../assets/google-logo.svg';
+import gitHubLogo from '../../../assets/gitHub-logo-black.svg';
+import googleLogo from '../../../assets/google-logo.svg';
 
 const OAuthButtons = () => (
   <>
